Rename `statemachine` -> `state-machine`.
It's more consistent with the other module names. (Yes, `typecheck` should
probably be renamed.)
