+++ local-modules/state-machine/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "statemachine",
+  "name": "state-machine",
   "version": "1.0.0",
   "main": "main.js",
